minor change to loader function to handle large values of norm resulting in arrays being automatically cast to 64 bit floats

+++ patchwork/_util.py
@@ -62,4 +62,4 @@ def _load_img(f, norm=255, num_channels=3, resize=None):
         img_arr = np.array(Image.fromarray(
                 (255*np.swapaxes(img_arr, 0,1)).astype(np.uint8)).resize(resize)
                     ).astype(np.float32)/norm
-    return img_arr
+    return img_arr.astype(np.float32)  # if norm is large, python recasts the array as float64
